requirements.txt
accelerate==1.1.0
fsspec==2024.9.0
pyarrow==18.0.0
loguru==0.7.2
numpy==1.26.4
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.5.2
torchmetrics==1.5.1
tqdm==4.66.6
# --find-links https://download.pytorch.org/whl/cu118
torch==2.5.0
faiss-gpu==1.9.0
gradio==5.7.1
onnx==1.16.0
redis==5.2.0
tensorrt==10.6.0.post1
protobuf==5.28.2